Move deploymetrics suite under metrics
